t/: Add test to get a TODO since the fork point
In order to include the 'bin.pm' file, we need to add its containing
folder to the include path. Do this by adding a Makefile with the
required command line argument.

Add a new test that modifies the test file to add a new new TODO line.
Test that it comes back in the `git todo` output.