write_since(): Take an arbitrary branch as input
Allow the branch to be parameterised instead of forcing the base to be
the local "master" branch.

Added the `Todos` struct to store a `Repository` reference, since we
need this for both the base ref and the diff.

Decided to take a `Tree` instead of a `Branch` because we might want to
compare against an arbitrary ref and allow users to specify that ref on
the command line.