write_since(): Output file paths relative to the current directory
Previously, the file paths of TODO lines in the output were relative to
the repository root. When you're in a subdirectory, though, this makes
them harder to work with, as they're relative paths, but aren't relative
to the current working directory.

Change the file path output so that paths are relative to the current
directory instead of the repo root.